refactor: tidy
